Updated delete player function
- now can't delete players during ongoing session
- matched dialogue box with the rest of the app

+++ app/src/main/java/com/mobdeve/machineproject/PlayerStatistics.kt
@@ -1,5 +1,6 @@
 package com.mobdeve.machineproject
 
+import android.app.AlertDialog
 import android.app.Dialog
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
@@ -11,6 +12,7 @@ import androidx.activity.result.ActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.mobdeve.machineproject.Model.GameSession
 import com.mobdeve.machineproject.Model.Player
 import com.mobdeve.machineproject.PlayerStatsRecycler.PlayerStatsAdapter
 import com.mobdeve.machineproject.SQL.DBHandler
@@ -26,21 +28,27 @@ class PlayerStatistics : ComponentActivity() {
     private lateinit var players: ArrayList<Player>
 
     private val displayConfirmPopup = {position: Int ->
-        val dialog = Dialog(this)
-        dialog.setContentView(R.layout.confirm_popup)
-        dialog.setCanceledOnTouchOutside(true)
-        dialog.show()
-
-        val yesButton = dialog.findViewById<Button>(R.id.confirmPopup_yes)
-        val cancelButton = dialog.findViewById<Button>(R.id.confirmPopup_cancel)
-
-        yesButton.setOnClickListener {
-            playerStatsAdapter.deletePlayerStat(position)
+        val confirmDialog = AlertDialog.Builder(this)
+        .setTitle("Confirm Player Deletion")
+        .setMessage("Are you sure you want to delete this player?")
+        .setPositiveButton("Yes") { _, _ ->
+            if (GameSession.players.isNotEmpty() && !GameSession.sessionConcluded){
+                AlertDialog.Builder(this)
+                .setTitle("Cannot Delete Player")
+                .setMessage("You cannot delete a player during an ongoing game session. Please finish or delete the current session first.")
+                .setPositiveButton("OK", null)
+                .create()
+                .show()
+            }
+            else{
+                playerStatsAdapter.deletePlayerStat(position)
+            }
+        }
+        .setNegativeButton("No") { dialog, _ ->
             dialog.dismiss()
         }
-        cancelButton.setOnClickListener {
-            dialog.dismiss()
-        }
+        .create()
+        confirmDialog.show()
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {
